src: refactor constants, deprecate require('constants')
The require('constants') module is currently undocumented and mashes
together unrelated constants. This deprecates the require('constants')
in favor of distinct os.constants, fs.constants, and crypto.constants
that are specific to the modules for which they are relevant. The
next step is to document those within the specific modules.